new binder.hpp takes classes IncludeSet, Binder from context.hpp
