+++ source/type.cpp
@@ -13,9 +13,9 @@
 
 #include <type.hpp>
 
-#include <options.hpp>
 #include <class.hpp>
 #include <enum.hpp>
+#include <options.hpp>
 #include <util.hpp>
 
 #include <clang/AST/ASTContext.h>
